+++ apps/bitcraft-local/src/server/game-data/bindings/global/index.ts
@@ -6925,3 +6925,4 @@ export class DbConnection extends __DbConnectionImpl<typeof REMOTE_MODULE> {
     return new SubscriptionBuilder(this);
   };
 }
+