+++ apps/bitcraft-local/src/server/game-data/bindings/regional/index.ts
@@ -7448,3 +7448,4 @@ export class DbConnection extends __DbConnectionImpl<typeof REMOTE_MODULE> {
     return new SubscriptionBuilder(this);
   };
 }
+