+++ apps/bitcraft-local/src/server/game-data/bindings/regional/types.ts
@@ -2909,6 +2909,9 @@ export const ExtractionRecipeDesc = __t.object("ExtractionRecipeDesc", {
   get spawnedPlaceables() {
     return __t.option(__t.array(ExtractionSpawnedPlaceable));
   },
+  get selfBuffs() {
+    return __t.option(__t.array(PlaceableSelfBuffChance));
+  },
 });
 export type ExtractionRecipeDesc = __Infer<typeof ExtractionRecipeDesc>;
 
@@ -2938,6 +2941,10 @@ export const FoodDesc = __t.object("FoodDesc", {
   consumableWhileInCombat: __t.bool(),
   get buffs() {
     return __t.array(BuffEffect);
+  },
+  autoConsume: __t.bool(),
+  get outputItemStacks() {
+    return __t.option(__t.array(ItemStack));
   },
 });
 export type FoodDesc = __Infer<typeof FoodDesc>;
@@ -6450,6 +6457,9 @@ export const ProspectingDesc = __t.object("ProspectingDesc", {
   },
   pctNodesForMaxContribution: __t.f32(),
   singleContributionOnly: __t.bool(),
+  get stepItemStacks() {
+    return __t.option(__t.array(ItemStack));
+  },
 });
 export type ProspectingDesc = __Infer<typeof ProspectingDesc>;